+++ tests/test_models.py
@@ -1,7 +1,7 @@
 import copy
 from datetime import datetime, timezone
 
-from pychan.models import Thread, Post, File
+from pychan.models import Thread, Post, File, Poster
 
 _TIMESTAMP = datetime.fromtimestamp(0, timezone.utc)
 
@@ -17,8 +17,10 @@ def test_thread_equality() -> None:
     # Test comparisons with non-Thread data types
     post = Post(t1, 1337, _TIMESTAMP, "test")
     file = File("test", "test")
+    poster = Poster("test", "test")
     assert t1 != post
     assert t1 != file
+    assert t1 != poster
 
 
 def test_post_equality() -> None:
@@ -33,8 +35,10 @@ def test_post_equality() -> None:
 
     # Test comparisons with non-Post data types
     file = File("test", "test")
+    poster = Poster("test", "test")
     assert p1 != t1
     assert p1 != file
+    assert p1 != poster
 
 
 def test_file_equality() -> None:
@@ -48,26 +52,50 @@ def test_file_equality() -> None:
     # Test comparisons with non-File data types
     thread = Thread("b", 1337)
     post = Post(thread, 1337, _TIMESTAMP, "test")
+    poster = Poster("test", "test")
     assert f1 != thread
     assert f1 != post
+    assert f1 != poster
+
+
+def test_poster_equality() -> None:
+    p1 = Poster("test", "test")
+    p2 = Poster("test", "test2")
+    p3 = Poster("test2", "test")
+    assert p1 == p1
+    assert p1 == p2
+    assert p1 != p3
+
+    # Test comparisons with non-Poster data types
+    thread = Thread("b", 1337)
+    post = Post(thread, 1337, _TIMESTAMP, "test")
+    file = File("test", "test")
+    assert p1 != thread
+    assert p1 != post
+    assert p1 != file
 
 
 def test_serialization() -> None:
+    file = File("test", "test")
+    assert tuple(file) == ("test", "test")
+
+    poster = Poster("test", "test")
+    assert tuple(poster) == ("test", "test")
+
     for text in [None, "test"]:
         thread = Thread("b", 1337, title=text)
-        assert tuple(thread) == ("b", 1337, text)
+        assert tuple(thread) == ("b", 1337, text, False, False)
 
-        file = File("test", text)
-        assert tuple(file) == ("test", text)
-
-        post = Post(thread, 1337, _TIMESTAMP, "test", poster_id=text)
-        assert tuple(post) == (
-            thread.board, thread.number, 1337, _TIMESTAMP, "test", False, text, None
+        post = Post(thread, 1337, _TIMESTAMP, "test")
+        # yapf: disable
+        assert tuple(post) == tuple(thread) + (
+            1337, _TIMESTAMP, "test", False, None, None, None, None
         )
+        # yapf: enable
 
-        post = Post(thread, 1337, _TIMESTAMP, "test", poster_id=text, file=file)
-        assert tuple(post) == (
-            thread.board, thread.number, 1337, _TIMESTAMP, "test", False, text, file.url
+        post = Post(thread, 1337, _TIMESTAMP, "test", file=file, poster=poster)
+        assert tuple(post) == tuple(thread) + (
+            1337, _TIMESTAMP, "test", False, file.url, file.name, poster.id, poster.flag
         )
 
 
@@ -79,22 +107,36 @@ def test_copying() -> None:
     assert thread is not thread_copy
     assert thread is not thread_deepcopy
 
-    file = File("test", name="test")
+    file = File("test", "test")
     file_copy = copy.copy(file)
     file_deepcopy = copy.deepcopy(file)
     assert file == file_copy == file_deepcopy
     assert file is not file_copy
     assert file is not file_deepcopy
 
-    post = Post(thread, 1337, _TIMESTAMP, "test", poster_id="test")
+    poster = Poster("test", "test")
+    poster_copy = copy.copy(poster)
+    poster_deepcopy = copy.deepcopy(poster)
+    assert poster == poster_copy == poster_deepcopy
+    assert poster is not poster_copy
+    assert poster is not poster_deepcopy
+
+    post = Post(thread, 1337, _TIMESTAMP, "test", file=file, poster=poster)
     post_copy = copy.copy(post)
     post_deepcopy = copy.deepcopy(post)
     assert post == post_copy == post_deepcopy
     assert post is not post_copy
     assert post is not post_deepcopy
+    # The assertions below test the behavior of copy versus deepcopy on the Post class
     assert post.thread == post_copy.thread == post_deepcopy.thread
     assert post.thread is post_copy.thread
     assert post.thread is not post_deepcopy.thread
     assert post.timestamp == post_copy.timestamp == post_deepcopy.timestamp
     assert post.timestamp is post_copy.timestamp
     assert post.timestamp is not post_deepcopy.timestamp
+    assert post.file == post_copy.file == post_deepcopy.file
+    assert post.file is post_copy.file
+    assert post.file is not post_deepcopy.file
+    assert post.poster == post_copy.poster == post_deepcopy.poster
+    assert post.poster is post_copy.poster
+    assert post.poster is not post_deepcopy.poster