+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Nothing currently!
 ### Added
 
 * The `Post` model now includes the post's timestamp in the `timestamp` field (by [@cooperwalbrun](https://github.com/cooperwalbrun))
+* The `Thread` model now has `stickied` and `closed` boolean fields (by [@cooperwalbrun](https://github.com/cooperwalbrun))
+* Created the `Poster` model and updated the `Post` model to use it (by [@cooperwalbrun](https://github.com/cooperwalbrun))
 
 ## v0.2.0 - 2022-07-29